Removed some warnings.  Improved Build scripts.

+++ mrViewer/src/core/iffImage.h
@@ -66,21 +66,21 @@ public:
     }
 
     virtual bool fetch( mrv::image_type_ptr& canvas,
-			const boost::int64_t frame );
+                        const boost::int64_t frame );
 
 protected:
     void end_read_chunk( FILE* f, iffChunk& chunk );
     void read_pixel_chunk( FILE* f, mrv::image_type_ptr& canvas,
-			   const int depth, const int bytes,
+                           const unsigned short depth, const int bytes,
                            const iffChunk& chunk );
 
-    void read_uncompressed_tile( FILE* file, 
-				 mrv::image_type_ptr& canvas,
-				 boost::uint8_t* data,
+    void read_uncompressed_tile( FILE* file,
+                                 mrv::image_type_ptr& canvas,
+                                 boost::uint8_t* data,
                                  const unsigned int compsize,
                                  const unsigned x, const unsigned y,
                                  const unsigned width, const unsigned height,
-                                 const short depth, const short bytes,
+                                 const unsigned short depth, const short bytes,
                                  const bool z);
     void read_chunk( FILE* f, iffChunk& chunk );
